+++ src/activities/home/HomeActivity.h
@@ -32,6 +32,7 @@ class HomeActivity final : public Activity {
   bool hasBookmarks = false;
   bool hasClippings = false;
   bool hasOpdsServers = false;
+  bool hasGrimmory = false;
   bool minimalMenuOpen = false;
   bool minimalSuppressInitialFrontRelease = false;
   bool homeBookSwapLongPressHandled = false;
@@ -74,6 +75,7 @@ class HomeActivity final : public Activity {
   void onSettingsOpen();
   void onFileTransferOpen();
   void onOpdsBrowserOpen();
+  void onGrimmoryBrowserOpen();
   void onReadingStatsOpen();
   void onSavedItemsOpen();